29.1.Loading Messages
Hmm. Refresh the page and watch closely. See it? When it first loads, the table is empty just for a moment. Then, when our AJAX call finishes, it gets filled in. That makes perfect sense, but it does make the page feel momentarily "broken"... and it may not always load this quickly.

No worries! This is why the Internet invented loading messages and animations! How can we add these in React? I have no idea! Let's find out!