29.2.Loading Messages
Here's our goal: before our AJAX call finishes, I want to render one row that just says "Loading". Hmm... this will need to happen inside RepLogList: that's where the tbody lives. But, hmm again: this component somehow needs to know whether or not the AJAX call has finished... and that is something that only RepLogApp knows.

To keep track of whether or not the AJAX call is finished, we need new state. On top, add some new state isLoaded: false.

Then, down below, when fetch() finishes, set isLoaded to true!

State, done! And thanks to how we're rendering RepLogs, this state is automatically pass as a prop. And now we start the prop-passing dance! In RepLogs, add the new prop type at the bottom: PropTypes.bool.isRequired. Oh, and you've probably noticed that I like to make pretty much everything required. That's a personal preference. Because this is my app, if I forget to pass a prop, it's probably a typo and I want to know.

Next, scroll up, destructure the isLoaded variable, find RepLogList, and pass that prop: isLoaded={isLoaded}.

Finally, do the same in that component: I'll steal the prop type and go up to destructure the variable.