29.3.Loading Messages
Ok, this is interesting: if the app is not loaded yet, we don't need to run any of this code down here. So, we can short-circuit the entire process: if !isLoaded, then return a completely new set of JSX, with a tbody, tr and <td colSpan="4" className="text-center">. Say, "Loading...".

Oh, and notice that this is colSpan with a capital "S". This is another, uncommon, case where the prop is slightly different than the HTML attribute. PhpStorm made it easy by auto-completing the correct version for React.

And... yea! That's it! Let's go refresh... but watch closely. There it was! And because React's model is so flexible, if you ever needed to, for some reason, reload the data, you could re-render that loading message simply by updating one piece of state. Nice.

We're on a roll! So let's make the delete link talk to our API.