32.2.Polyfills: fetch & Promise{}
Super easy! To make sure fetch() is available everywhere, we can import it from our entry file: rep_log_react.js. Then, it will definitely be available in rep_log_api.js. But... I like to get even crazier! Open layout.js. Then look inside webpack.config.js. layout is configured as my "shared" entry... which is a fancy way of saying that layout.js is included on every page and so its code will always run. Inside that file, import 'whatwg-fetch'.

Every part of my app can now safely rely on the fact that the global fetch() function will be available.

Oh, and because I love digging in to see how things work, let's go check out the polyfill code! Open node_modules, search for whatwg, expand its directory and open fetch.js. This is the file we just imported.

Look all the way at the bottom: it's a self-executing function. It passes this into the function, which in a browser environment, is the global window object. Then, on top, that window object is passed as a self. And because all global variables & functions are actually properties on the window object, it's able to ask:

Hey! Does the window variable have the global fetch() function on it?

If it does, it just returns. If it does not, the rest of this code works to define and add it. There it is: self.fetch =. This is a polyfill in action - kinda cool.