32.3.Polyfills: fetch & Promise{yarn add promise-polyfill --dev}
The Promise Polyfill

Go back to the fetch polyfill docs. Ah, it says that we also need to use a Promise polyfill for older browsers. We can literally see this inside of the fetch polyfill: it assumes that a Promise class is available.

Let's polyfill it to be safe: click into the library they recommend. Cool: copy the name, move over, and:

yarn add promise-polyfill --dev