34.2.Temporary Messages & componentWillUnmount{Clearing the Message with setTimeout()}
Clearing the Message with setTimeout()

Back in setSuccessMessage(), to clear the message, use setTimeout(), pass it an arrow function, and use this.setState() to reset successMessage back to empty quotes. Let's do that after 3 seconds.

Ok! Let's give this a try! Refresh and lift my big fat cat 5 times. Success! And... gone!

So easy! Except... yes... it was too easy. There's an edge case: if setSuccessMessage() is called once, then called again 2 seconds later, the second message will disappear after only 1 second! It's a minor detail, but we can code this better.