34.3.Temporary Messages & componentWillUnmount{Clearing the Timeout}
Clearing the Timeout

Basically, before we call setTimeout, we want to make sure to clear any previous timeout that may be waiting to fire. The setTimeout() function returns an integer, which we can use to clear it. To keep track of that value, in the constructor, initialize a new property: this.successMessageTimeoutHandle = 0.

This has nothing to do with React: we're just taking advantage of our object to store some data. Oh, and the value 0 is just a "null" value in disguise: if we pass this to clearTimeout(), nothing will happen.

Back down in setSuccessMessage, before setTimeout, add clearTimeout(this.successMessageTimeoutHandle). To set that property, add this.successMessageTimeoutHandle = before setTimeout().

And finally, to be completely on top of things, inside the callback, after we reset the state, set the timeout handle back to 0.

And... we're done! Our message will always disappear a full 3 seconds after the last success message has been set. Except... yea... there is still one teenie, tiny problem... and this time, it's special to React.