35.3.Updating Deep State Data{Success Message on Delete}
Success Message on Delete

We have this cool system where we can lift our cat 26 times and see this temporary success message. So, we might as well do the same thing when we delete. And this is easy! Inside of RepLogApp, down in handleDeleteRepLog, chain off the delete: .then(), an arrow function, and this.setSuccessMessage(): Item was Un-lifted.

Cool! Move back and try it! Success... message.