35.4.Updating Deep State Data{"Ghosting" the Deleting Row}
"Ghosting" the Deleting Row

We could be satisfied with our loading & success message setup. But... if you want... we can get fancier! Before this code change, we deleted the rep log state immediately, but we didn't show the success message until after the AJAX call finishes. If you want that to feel more synchronized, we could move the setState() call so that it fires only when the rep log is actually deleted.

But, we're trading problems. Refresh again. When you click delete, there's a slight pause before the user gets any feedback. I'll add a few more items to the list real quick so that we can keep deleting.