36.2.Server Validation & fetch Failing{Making fetch Fail}
Making fetch Fail

This behavior... isn't great. So, I like to fix it: I'll paste in a new function called checkStatus(). If we call this function and the status code is not 200 or 300, it creates an Error object, puts the response onto it, and throws it. By the way, you could change this logic to also throw an error for 300-level status codes, that's actually how jQuery works.

To use this, back up in fetchJson(), add this handler: .then(checkStatus).

Let's try it! Refresh, select our bad item, a number and... yes! This is a much more obvious message:

  Uncaught (in promise) Error: Bad Request at checkStatus

Now that fetch is behaving better, let's use this error response to add a message for the user.